[Silver II] Title: 최대 곱, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1500. 최대 곱/최대 곱.cc
@@ -7,8 +7,8 @@ int main() {
     
     cin >> S >> K;
 
-    long long q = S/K;
-    long long sum = q*K;
+    int q = S/K;
+    int sum = q*K;
     long long res = 1;
 
     for (int i = 0; i < S - sum; i++) res *= (q + 1);
